Combine header and search form
This change gives more space to definitions and examples, which is
particularly important on smaller screens. A logo is substituted for
the text header in the main view, and the text header is re-shown
(without the search form) in menu and stats views.
TODO: is the logo right, or would an abbreviated text header be
better?