Move jieba off the main thread
This prevents UI jank while typing, as previously tokenization was
happening on the main thread. This does increase search suggestion
worker memory, however, and upcoming commits will need to find a
way to decrease that.

The incompatibility between CEDICT and jieba/Intl stuff remains a
TODO (especially with the nasty vetCandidate function).