Fix login header; make (minimal) benefits clear
The prior commit's continued abuse of CSS grid caused more wrapping
than was previously present (and it was already bad).

It's also good to make clear to users that registration is not
required, and that its only benefits are with syncing across devices.
(which also keeps in line with the static website spirit of the app).