Only special case empty math results due to subtraction
Previously, operations like 我+你+我 would return results.
This was because empty results of prior operations effectively
reset the result, but we only want that with subtraction.
我+你+我 now returns [].

Note that this does not address self-adds, which are still broken.
It also continues adding all parents in cases like: 我-我+我.
The same transitivity is true for normal operations (so 力+口
returns many characters with 另 or 加 as a component, intentionally).