Hack around android keyboard oddities
It's unclear if this occurs on all android versions and all
browsers vs specific chrome versions or something.
But by default, the submit preventDefault seems to not cause the
input to trigger a blur, and the on-screen keyboard triggers
window resize events. Ideally this change is safe for other
platforms where this isn't occurring as well.