Hack around soft keyboard oddities (#53)
It's unclear if this occurs on all android versions and all
browsers vs specific chrome versions or something, but
the on-screen keyboard triggers window resize events. 
Because the graph is re-rendered on resize, you get an 
annoying flicker. This does not happen on iOS.

Ideally this change is safe for all platforms.