Add popover menu, replacing add-all pattern (#67)
Previously, there was an add button by the word header, which
would add all flash cards. This was convenient if you like all
the sentences, but if not, it quickly results in cruft, and
if you have a bunch due, it's hard to remember which you wanted
to delete vs keep.

This also sets us up for adding more buttons.
* Copy to clipboard
* this word in N cards
* generate more sentences via AI API
* explain grammar via AI
etc.