Switch to SM2 SRS algorithm (#84)
FSRS is still on the roadmap, but this is a quick and easy first
step, while FSRS will be a bit more involved and include additional
dependencies. There are various other flashcard features coming
as well, like showing flashcard users how many words or characters
are not in their study list beside example sentences.