Allow ease to grow with correct answers (#87)
Still should use FSRS instead...
But the prior implementation caused ease to only shrink or stay constant.
Now, ease grows when the user says they got the right answer.
Another option would be to add a button for "got it right, but it was hard",
but I prefer yes/no. It's less important that a user memorize each sentence
anyway, it's more about consistent exposure to plenty of Chinese text, with
varied vocabulary.