Don't strip punctuation for multi-word or AI searches (#88)
These punctuation marks should actually be treated like any other
non-word token in multi-word searches, and they should be included
in text sent back to the AI for explanation.

The display of non-word tokens in multi-word searches should probably
be made less prominent, but this commit leaves that change for another
time.