Fix duplicate word tone colors (#90)
This allows tone colors to work for duplicate words in a sentence.
The previous code was erroneously using an ID instead of a class.
This is rarely an issue, but occurs if a user searches for text with
duplicate words outside the top N for which pinyin are not loaded up
front.