Add AI sentence generation feature
This feature allows eligible users to generate AI sentences for
a word. Each definition of the word is passed in, and the model
is instructed to make sentences for each definition, so that
learners can see a variety of uses. There are some TODOs, for
example some definitions don't make sense to generate sentences for,
but this is a good start.

Coming soon is also collocation analysis and example sentence
generation, to be integrated on the flow tab.