Add support for gamepad controls on flash cards
This allows you to connect a controller (e.g., switch pro) and use it
to perform flash card actions. A/B (since xbox and switch differ) flip
the card, Y/X plays the audio, and R/L (RB/LB) mark right or wrong,
respectively.

This is quite silly but why not. It's kinda nice to do flashcards
this way. This change also adds a keyboard shortcut to play audio.

It's rather unfortunate that the web platform doesn't have button
events for gamepads like for keyboards, but I tried to mimic it
with a simple state machine and custom events.