Add color coding for flashcard suitability (#119)
This change adds a heuristic for determining how good a flashcard a
given sentence is likely to be. Sentences where:
* most of the words are already in your flashcard set
* those that aren't already being studied are high-frequency words
are usually better flashcards in that they are closer to comprehensible
input and the new words are high value. This heuristic can be refined
over time.

Another TODO is to let users define their own priority based on frequency
ranks. For now, the range 0 - 10k is used. One can also imagine a different
graph color coding mechanism that emphasizes high priority words.